PythonApplications/CLIVideo/pyproject.toml
[tool.poetry]
name = "clivideo"
version = "0.1.0"
description = "Command Line Interface for Video Generation"
authors = ["In Service of X LLC <inserviceofxy@proton.me>"]
readme = "README.md"
packages = [
    { include = "clivideo" },
    { include = "Executables" }
]

[tool.poetry.dependencies]
python = "^3.10"
prompt-toolkit = "^3.0.50"
wcwidth = "^0.2.13"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
